+++ src/main/java/org/zalando/nakadi/service/subscription/SubscriptionService.java
@@ -188,7 +188,9 @@ public class SubscriptionService {
     public Result<Void> deleteSubscription(final String subscriptionId, final Client client) {
         try {
             final Subscription subscription = subscriptionRepository.getSubscription(subscriptionId);
-            client.checkId(subscription.getOwningApplication());
+            if (!client.idMatches(subscription.getOwningApplication())) {
+                return Result.forbidden("You don't have access to this subscription");
+            }
 
             subscriptionRepository.deleteSubscription(subscriptionId);
